+++ src/main/java/ru/prikhodko/repository/LanguageRepository.java
@@ -3,5 +3,5 @@ package ru.prikhodko.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.prikhodko.vo.Language;
 
-public interface LanguageRepository extends JpaRepository<Language, Long> {
+public interface LanguageRepository extends JpaRepository<Language, Integer> {
 }
